+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Left-click on rares to open up the group finder searching for them by name (thanks znf)
 * Many more treasures in Argus (thanks znf)
+* Missing from Azsuna
 * You know what, Broken Shore too. Bit out of order, but there we go.
 
 ## Changed in v12